web: preparation for patternfly 5 upgrade
This commit doesn't change anything at all inside the UI.  It's just re-arranging a few details in
the way styles are composed into our Web Components.

- The return type for `get styles()` has been removed in favor of type inference.  More to the
  point, if the return type *doesn't* conform to Lit's expectation, type checking still complains
  quite effectively.
- All in-line custom CSS has been moved out and above the class declaration, given a standard
  variable name (`customCSS`) whenever possible, and the variable placed into the style array in its
  place.
- Declarations of the form `super.styles.concat(...)` have been replaced with `[...super.styles, ...]`.

The purpose of this is simple: to establish and maintain a *consistent* and *predictable* pattern
that our refactoring tools can use.  It also allows us to survey our uses and see where there are
intriguing patterns that we could exploit to make our life easier.

For example, the UserSetting page has so much going on, it seems almost as if we would benefit from
thinking harder about *why* there are so many different kinds of things in this object.

```
return [PFBase, PFPage, PFDisplay, PFGallery, PFContent, PFCard, PFDescriptionList, PFSizing, PFForm, PFFormControl, PFStack, customCSS];
```

On the other foot, we have these which appear *multiple* times in our codebase, again leading one to
wonder what these 4, 5, 7, or 10(!) pages have in common and how we could extract what's common in
them out and deduplicate that effort.

```
 4         return [PFBase, PFLogin, PFForm, PFFormControl, PFButton, PFTitle];
 4         return [PFBase, PFLogin, PFForm, PFFormControl, PFTitle, PFButton];
 5         return [PFBase, PFForm, PFButton, PFRadio];
 7         return [PFBase, PFButton, PFRadio];
10         return [...super.styles, PFDescriptionList];
```

That last one is especially intriguing because there are 36 (!) `DescriptionLists` in our code base.
Creating a declarative variant of it with its own styles embedded would free up a lot of the CSS
hand-work.