web: revise CSS handling for esbuild
Replace some ad-hoc css mangling with the `unsafeCSS()` function as recommended
by Lit.  This also streamlines a bit of the CSS handling so that it's not calling
`ensure` multiple times.

The worst part of this is understanding just how much of this mangling is really
here to keep Storybook happy.