web: improve build speeds even moar!!!!!! (#8954)
* web: fix esbuild issue with style sheets

Getting ESBuild, Lit, and Storybook to all agree on how to read and parse stylesheets is a serious
pain. This fix better identifies the value types (instances) being passed from various sources in
the repo to the three *different* kinds of style processors we're using (the native one, the
polyfill one, and whatever the heck Storybook does internally).

Falling back to using older CSS instantiating techniques one era at a time seems to do the trick.
It's ugly, but in the face of the aggressive styling we use to avoid Flashes of Unstyled Content
(FLoUC), it's the logic with which we're left.

In standard mode, the following warning appears on the console when running a Flow:

```
Autofocus processing was blocked because a document already has a focused element.
```

In compatibility mode, the following **error** appears on the console when running a Flow:

```
crawler-inject.js:1106 Uncaught TypeError: Failed to execute 'observe' on 'MutationObserver': parameter 1 is not of type 'Node'.
    at initDomMutationObservers (crawler-inject.js:1106:18)
    at crawler-inject.js:1114:24
    at Array.forEach (<anonymous>)
    at initDomMutationObservers (crawler-inject.js:1114:10)
    at crawler-inject.js:1549:1
initDomMutationObservers @ crawler-inject.js:1106
(anonymous) @ crawler-inject.js:1114
initDomMutationObservers @ crawler-inject.js:1114
(anonymous) @ crawler-inject.js:1549
```

Despite this error, nothing seems to be broken and flows work as anticipated.

* web: improve build speeds even moar!!!!!!

While investigating how to improve the integration of Patternfly 5
into our product, I came across a hint on how to pre-process the
stylesheets into CSSStylesheetObjects on the fly. While trying to
integrate that hint into our own build process, I got an error
message about how esbuild plugins can't be used with the synchronous
API yet.

So, being even more curious, I tried to figure out how to make our
multiple builds work with the asynchronous API.

Then I wondered how it behaved with `Promise.allSettled().`

The result is a build time of less than one second.

Can't complain.

* web: moar speed plz!!!

- Re-arrange the build order so the larger components get built first
- Change the criteria for "what is a proxy object."
- Adds some (probably trivial) awaits() where expected.

* add comment for ordering

Signed-off-by: Jens Langhammer <jens@goauthentik.io>

---------

Signed-off-by: Jens Langhammer <jens@goauthentik.io>
Co-authored-by: Jens Langhammer <jens@goauthentik.io>