web: move context controllers into reactive controller plugins
While I was working on the Patternfly 5 thing, I found myself cleaning up the
way our context controllers are plugged into the Interfaces.  I realized a
couple of things that had bothered me before:

1. It does not matter where the context controller lives so long as the context
   controller has a references to the LitElement that hosts it.
   ReactiveControllers provide that reference.
2. ReactiveControllers are a perfect place to hide some of these details, so
   that they don't have to clutter up our Interface declaration.
3. The ReactiveController `hostConnected()/hostDisconnected()` lifecycle is a
   much better place to hook up our EVENT_REFRESH events to the contexts and
   controllers that care about them than some random place in the loader cycle.
4. It's much easier to detect and control when an external change to a
   context's state object, which is supposed to be a mirror of the context,
   changes outside the controller, by using the `hostUpdate()` method.  When the
   controller causes a state change, the states will be the same, allowing us to
   short out the potential infinite loop.

This commit also uses the symbol-as-property-name trick to guarantee the privacy
of some fields that should truly be private. They're unfindable and
inaddressible from the outside world. This is preferable to using the Private
Member syntax (the `#` prefix) because Babel, TypeScript, and ESBuild all use an
underlying registry of private names that "do not have good performance
characteristics if you create many instances of classes with private fields"
[ESBuild Caveats](https://esbuild.github.io/content-types/#javascript-caveats).