web:fix markdown rendering bug for alerts
The move to using showdown dynamically, at run-time, resulted in a parse error
where our alerts were not being decorated with the right syntax. This patch
recognizes the new `:::info` EOL syntax (and leaves the old one in-place, as
well) and the rendering is now correct.

Our complexity has reached the point where eslint now needs the memory increase.